Clean up MaterializeTablesSuite and error on partitioning changes (#118)
Per the discussion in [this Slack
thread](https://databricks.slack.com/archives/GPZMR1L7J/p1747930783113339),
raise an error if the declared partition cols for a table or MV don't
match the partition cols on the actual table. This gives us maximum
flexibility on how to handle this in the future (e.g. automatically
evolve the cols or leave them unchanged).

Also add a bunch of tests (including for the above behavior) to
`MaterializeTablesSuite`.

After this PR, `MaterializeTablesSuite` should be ready for public
consumption.

### How I tested

Ran the CI script