Put spark-submit underneath the hood of the spark-pipelines CLI (#126)
Updates our `spark-pipelines` CLI script to use `spark-submit`, as
recommended by @martin-grund_data.

This will enable running `spark-pipelines` against common cluster
managers.

A nice consequence of this change is that users will be able to avoid
launching a separate Spark Connect server and pointing to it via the
`--remote` arg. We follow the default behavior of `spark-submit` with
Spark Connect, where Spark handles launching the server for the user and
tearing it down at the end of the process. So `spark-pipelines run` will
just work if there's a pipeline spec in the directory.